requirements.txt
streamlit>=1.25
pandas>=1.5
numpy>=1.23.5
scikit-learn>=1.2.2
joblib>=1.3

psycopg2-binary>=2.9

boto3>=1.28

google-cloud-aiplatform>=1.38

python-dotenv>=1.0